price to int

+++ src/main/java/com/ohio/grand_hotel_ohio/dto/RoomDTO.java
@@ -4,7 +4,6 @@ package com.ohio.grand_hotel_ohio.dto;
 import com.fasterxml.jackson.annotation.JsonInclude;
 import lombok.Data;
 
-import java.math.BigDecimal;
 
 import java.util.List;
 
@@ -14,13 +13,13 @@ public class RoomDTO {
 
     private Long id;
     private String roomType;
-    private BigDecimal roomPrice;
+    private int roomPrice;
     private String roomPhotoUrl;
     private String roomDescription;
 //    private List<BookingDTO> bookings;
 
 
-    public RoomDTO(Long id, String roomType, BigDecimal roomPrice, String roomPhotoUrl, String roomDescription) {
+    public RoomDTO(Long id, String roomType, int roomPrice, String roomPhotoUrl, String roomDescription) {
         this.id = id;
         this.roomType = roomType;
         this.roomPrice = roomPrice;
